enable resharding v2 (#10303)
# Feature to stabilize
This PR stabilizes the Resharding V2 protocol feature. It introduces a
new implementation for resharding and a new shard layout for the
production networks - mainnet and testnet. The resharding process from
the current shard layout to the new shard layout will be triggered
automatically an epoch before the switch to the new protocol version.

# Context

- https://github.com/near/NEPs/pull/508
- https://github.com/near/nearcore/issues/8992

# Testing and QA

This feature was extensively tested in unit tests, integration tests,
nayduck tests, mocknet, testnet and mainnet.

# Checklist
- [ ] nayduck - https://nayduck.near.org/#/run/3308 
- [x] Update CHANGELOG.md to include this protocol feature in the
`Unreleased` section.