feat: Change default shard trie cache size to 500MB (#10373)
Current value is 50MB which might not be enough when the shard is
overloaded. One example when this can be an issue is when delayed
receipts queue grows to 20k+ items, which already happened during NEAT
inscription load. Insufficient cache size results in increased number of
disk reads to State rocksdb column. That increases chunk application
time which results in more missed chunks.
This only applies to shards 0,1,2 since we have an overwrite for shard
3. So overall expected RAM increase is 3 * (500 - 50) = 1350MB.