[stateless validation] Real ChunkStateWitness struct, and prototype validation logic (#10381)
This fills in the ChunkStateWitness struct with real data that would be
included in a chunk state witness.

It also includes a prototype implementation for validating the
ChunkStateWitness against the chain, which includes a pre-validation
part where the ChainStore is used to retrieve necessary blocks and chunk
headers (to verify receipts and transactions against), as well as the
main validation part where transactions and receipts are applied with
the runtime.

This verification logic is untested right now because the state witness
production part is not implemented.

#9292