fix(split-storage) - optimize copying of state (#10403)
The current approach means that when looking for a trie node in the db
we need to check all shards. In practice most nodes are only ever
present in one shard so the current implementation is ~4x slower when it
comes to reading trie nodes than it could be. The state reads also
happen to account for majority of time spent in copying heights from hot
to cold.

This fix should resolve the ever growing cold head lag that is described
in more detail on
[zulip](https://near.zulipchat.com/#narrow/stream/308695-pagoda.2Fprivate/topic/Split.20storage.20nodes.20increasing.20cold.20head.20lag)
which is a major issue for our archival nodes.