style: RFC to enable clippy::perf lints (#10378)
During the holidays I took notice of a discussion about enabling clippy
against the `rust-lang/rust` repository. Within that discussion people
also debated which lints or lint categories they would consider
important to enable and to my surprise `clippy::perf` was one of those
categories.

So I went ahead and checked out what it entails for our nearcore
codebase as well. In broad strokes a large number of lint triggers were
in test code. Those changes aren’t particularly interesting – test code
is not particularly perf sensitive. But in certain cases it has also
resulted in cleaner code (like for instance `except(format!(...))` is a
code smell IMO.)

For regular code the most notable class of changes was about the size of
data structures. For instance, we had a couple of cases where a return
value would be a result of a small successful value and an unreasonably
large error variant. When functions like these are called frequently,
moving around the return value can become pretty darn expensive.

In many cases the lints had useful suggestions that I could apply
immediately without having to think what a better way to write down the
same code would be, so it doesn’t seem like it would be too harmful to
the developer productivity...