feat(resharding) - delete snapshot immediately after resharding is finished (#10450)
As long as the snapshot is present the storage overhead will keep
increasing. This PR makes it so that the snapshot is deleted immediately
after resharding is finished. The snapshot is only deleted if the node
is not configured to take snapshots every epoch.

Initially I wanted to put the trigger logic somewhere between sync jobs
actor and chain but there it was hard to tell if resharding is finished
for **all** the shards. It was made trickier by the fact that we only
resharding shards that we will track next epoch and should only check
those. At the end of the day I placed it in the state sync logic where
it'll be triggered once all state syncs and hence all reshardings are
done.