fix codecov intermittent failure on PRs from forked repos (#10470)
Forked repos do not actually have access to the codecov token, and
codecov is subject to rate-limiting which leads to intermittent
failures.

Making the token public is not a big deal, because we really treat all
information on codecov as informational only, and do not publicize it as
an important part of our system.

However, it’s likely that someone will report a bug bounty about it, so
I also added a quite descriptive explanation of the fact we do not care.

The alternatives are:
- Live with the intermittent failure from codecov uploading
- Completely get rid of codecov, and switch to using only llvm-cov
everywhere — but this’ll lose us the nice red checkmark when coverage is
not good