install tools from github releases rather than building and caching them (#10460)
This should help us with the number of caches for the tools we use. Each
one is lightweight, but they are a lot.

Currently, 23 of the 25 caches on the github first page are
`cargo-install` caches, and that just continues in the next pages. My
guess is the action doesn’t deal with the cache properly and has the
same cache uploaded multiple times, because we don’t even have that many
cargo-install calls. Unfortunately, reading the code of the
cargo-install action, I was not able to see an obvious reason why it
resulted in this.