feat: add shadow stateless validation (#10478)
This is the first PR for #10440.

New feature `shadow_chunk_validation` is introduced. When enabled it
makes the node produce state witness for every chunk in every processed
block and then self-validate it. New metrics are introduced for
validation failures and state witness size.

Detailed state witness size breakdown metrics from
[robin/witness-size](https://github.com/near/nearcore/compare/robin/witness-size)
will be added in a separate PR.