feat(state-sync): Download state header from external storage (#10515)
In a previous PR, we have introduced the option to upload the state
parts to external storage.
In this PR we are changing the way that state sync headers are
requested, introducing to download
from external storage. The motivation behind this is that sometimes
peers may not track all shards
making it impossible for them to share the state sync header.

The most important change is in commit #3 

I had an rpc node running on mainnet and testnet and it was able to
sync. [Grafana](https://nearinc.grafana.net/goto/3lNKbppSg?orgId=1) The
long testnet sync header download was because there was not node to
upload the state in the external bucket. Node started downloading as
soon as I started the uploading node.