Add a way to wait until a state witness is fully processed, which makes writing integration tests easier (#10559)
State witnesses are processed asynchronously - a task is spawned on a
separate thread, which validates the state witness and then sends out
chunk endorsements. Currently there's no way to directly wait for this
processing to finish, one can only wait for the chunk endorsements to
appear. This makes writing integration tests difficult. Let's add a way
to wait for the processing to finish, which will make writing the tests
easier.

Details:
There's no network stack in integration tests, all messages are passed
around manually. We first propagate all state witness messages, which
starts their processing in the background, but then we have to wait for
this processing to finish before propagating chunk endorsements. As
there's no way to wait for it, we have to do tricky things. The current
approach is to
define the set of expected chunk endorsements that should appear, and
then wait for the right endorsement messages to be sent out. This
approach has some limitations - it's often hard to figure out which
endorsements should be sent out exactly, as the tests create various
strange scenarios - missing chunks, blocks, forks, etc... It requires a
lot of manual work to fully understand the test and prepare the right
set. It's also quite fragile, any change in chunk endorsements (e.g not
sending them to ourselves) will break all the tests.
Adding the ability to wait for witness processing to finish greatly
simplifies things. We can just pass around witnesses, wait for them to
finish, and then propagate the generated endorsements. Without defining
any expected sets to wait for.
As an example of the benefits offered by this solution, compare the fix
in https://github.com/near/nearcore/pull/10537 to the one added in this
PR. The one in https://github.com/near/nearcore/pull/10537 is much more
complex and fragile.

I saw that there is also a thing called `ChunkEndorsementTracker`, which
tracks what chunk endorsements were sent out, but IMO it's a different
thing from `ProcessingDoneTracker`. We were already able to track which
endorsements are sent out in the tests, but that didn't make writing
tests easier. Additionally, in some tests we send out invalid chunks,
which wouldn't produce any endorsements at all, so we couldn't wait for
them.

Refs: https://github.com/near/nearcore/issues/10506,
https://github.com/near/nearcore/pull/10537, [zulip
conversation](https://near.zulipchat.com/#narrow/stream/407237-pagoda.2Fcore.2Fstateless-validation/topic/A.20way.20to.20wait.20for.20state.20witness.20processing.20to.20finish/near/419062119)