fix: make adversarial_behaviors tests pass with stateless validation (#10571)
## Description
The tests in `adversarial_behaviors.rs` didn't work with stateless
validation, let's make them pass again.
To achieve this we first need to propagate chunk state witnesses and
endorsements, and then adjust the expected behavior to match the one
exhibited by stateless validation.

The most complex part of this PR is modifying which blocks should be
skipped in
`test_banning_chunk_producer_when_seeing_invalid_chunk_base()`.
This function creates a blockchain with one malicious chunk producer and
runs it for a few epochs.
Which blocks will be skipped depends on whether we are using stateless
validation or not.

### Without stateless validation
In the old version of the protocol the invalid chunk is included in a
block and then the whole block is validated. The chunk is invalid, so
the whole block is also invalid and the block should be skipped. After
that the malicious chunk producer is banned for the rest of the epoch,
which means that block producers will reject chunks sent by this chunk
producer, so the subsequent blocks won't contain any invalid chunks and
won't be skipped.
Only the first block with invalid chunks in an epoch should be skipped. 

### With stateless validation
With stateless validation the situation is entirely different. When the
malicious chunk producer produces an invalid chunk, chunk validators
will refuse to send out endorsements for this chunk, so the chunk won't
be included in any block. This means that no blocks should be skipped.

There is one exception - in the test the block produced at height 2 is
invalid and gets skipped. I suspect that this is because the first few
blocks after genesis don't use stateless validation, and they are still
handled by old protocol. I'm not 100% sure if that's the case,
@Longarithm could you confirm that this explanation makes sense?
Something is definitely different for the first few blocks, in the logs
I can see `client: Banning chunk producer for producing invalid` at
height 2, which seems to belong to the old protocol. In the new version
malicious chunk producers are banned using `BanPeer`, without any log
messages.

Refs: https://github.com/near/nearcore/issues/10506, [zulip
discussion](https://near.zulipchat.com/#narrow/stream/295558-pagoda.2Fcore/topic/Skipping.20the.20first.20block.20with.20missing.20chunks.20in.20an.20epoch.3F/near/419838610)