fix: make test_bad_chunk_mask work with stateless validation (#10583)
This test didn't work with stateless validation because it didn't
propagate chunk state witnesses and endorsements as needed.

I modified the test to use `TestEnv`, which allowed me to use the
exisiting functionality for propagating all the partial chunk requests,
state witnesses and endorsements.
This changed the assignment of chunk and block producers. Previously one
of the clients was a chunk producer and the other a block producer, with
their roles swapped on each block. Now the roles are assigned based on
the logic in `EpochManager`. IMO it doesn't affect the test, in the
crucial moment when the bad chunk mask is tested, the chunk producer is
different from the block producer, just like in the original test.

The test manually produces a single chunk on shard 0 to achieve the
desired chunk mask, so I couldn't just use the default way of generating
chunks during block processing. I had to implement a function that would
manually produce the required chunk.

Refs: https://github.com/near/nearcore/issues/10506